requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.1.3
execnet==2.0.2
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.2.0
py==1.11.0
PyMySQL==1.1.0
pytest==7.4.3
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.4.0
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.7
xlrd==1.2.0
zipp==3.15.0
